requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
loguru==0.5.3
multidict==4.7.6
mysql-connector-python==8.0.21
protobuf==3.12.2
pyTelegramBotAPI==3.7.2
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
telebot==0.0.3
urllib3==1.25.9
yarl==1.4.2
